+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.8.0...HEAD)
+## [Unreleased](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.9.0...HEAD)
+
+## [1.9.0](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.8.0...v1.9.0)
+
+### Added
+Fields
+- `client_cert` (String) Client certificate.
+- `client_private_key` (String) Client private key
+to `fivetran_proxy_agent` resource
+
+### Removed
+Fields `salt` and `proxy_server_uri` from `fivetran_proxy_agent` resource
+Field `salt`, `token` from `fivetran_proxy_agent` datasource.
+Field `salt` from `fivetran_proxy_agents` datasource.
 
 ## [1.8.0](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.7.0...v1.8.0)